fix #7141 -- when resuming a subscription show the correct cost and properly enable payment flow

+++ src/packages/server/purchases/resume-subscription.test.ts
@@ -9,10 +9,13 @@ import getPool, { initEphemeralDatabase } from "@cocalc/database/pool";
 import { uuid } from "@cocalc/util/misc";
 import { createTestAccount, createTestSubscription } from "./test-data";
 import dayjs from "dayjs";
-import resumeSubscription from "./resume-subscription";
+import resumeSubscription, {
+  costToResumeSubscription,
+} from "./resume-subscription";
 import cancelSubscription from "./cancel-subscription";
 import { getSubscription } from "./renew-subscription";
 import getLicense from "@cocalc/server/licenses/get-license";
+import getBalance from "./get-balance";
 
 beforeAll(async () => {
   await initEphemeralDatabase({});
@@ -28,9 +31,8 @@ describe("create a subscription, cancel it, then resume it", () => {
   let license_id = "";
   it("creates an account, license and subscription", async () => {
     await createTestAccount(account_id);
-    ({ subscription_id, license_id } = await createTestSubscription(
-      account_id
-    ));
+    ({ subscription_id, license_id } =
+      await createTestSubscription(account_id));
   });
 
   it("tries to resume the subscription we just made which isn't canceled and get error, since only canceled subscriptions can be resumed", async () => {
@@ -53,10 +55,19 @@ describe("create a subscription, cancel it, then resume it", () => {
     // fully refunded (since starts in future) -- license is not active for nonzero period
     expect(license.expires).toBe(license.activates);
 
+    const balanceBeforeResume = await getBalance(account_id);
+    const costToResume = await costToResumeSubscription(subscription_id);
     // now resume:
     await resumeSubscription({ account_id, subscription_id });
     // and it is active
     expect((await getSubscription(subscription_id)).status).toBe("active");
+
+    // and balance is right
+    const balanceAfterResume = await getBalance(account_id);
+    expect(balanceBeforeResume - costToResume).toBeCloseTo(
+      balanceAfterResume,
+      2,
+    );
 
     // confirm the license is active again for same period as subscription current period.
     const license2 = await getLicense(license_id);
@@ -72,9 +83,7 @@ describe("create a subscription, cancel it, then resume it", () => {
       cancelImmediately: true,
     });
     const pool = getPool();
-    await pool.query("DELETE FROM purchases WHERE account_id=$1", [
-      account_id,
-    ]);
+    await pool.query("DELETE FROM purchases WHERE account_id=$1", [account_id]);
     expect.assertions(1);
     try {
       await resumeSubscription({ account_id, subscription_id });
